requirements.txt
numpy>=1.24,<2.0
pandas>=2.0,<3.0
scipy>=1.11,<2.0
scikit-learn>=1.3,<2.0
statsmodels>=0.14,<1.0
pyarrow>=14.0,<18.0
pyyaml>=6.0,<7.0
matplotlib>=3.8,<4.0
seaborn>=0.13,<1.0
networkx>=3.1,<4.0
tqdm>=4.66,<5.0
rpy2>=3.5,<4.0
pytest>=7.4,<9.0
jupyter>=1.0,<2.0
nbformat>=5.9,<6.0
